requirements.txt
gradio~=3.43.2
ultralytics
pandas~=2.1.0
requests~=2.31.0
torch~=2.0.1
Pillow~=10.0.0
torchvision~=0.15.2